requirements.txt
pyyaml
docker
setuptools
wheel
boto3
halo
